backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
groq
requests
